refactor(tools): extract demo seed subsystem
Moved the demo seed and cleanup cluster from firstslice.go into demo_seed.go without changing symbols or behavior. Gates: go build ./internal/tools, go build ./..., go vet ./..., go test -count=1 ./..., golangci-lint run, make catalog-drift.
